Hinge Locking system Refactor
