+++ Assets/Scripts/Interaction/HingeLocker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e1b559fe634ae3c4eae322a79d587e2c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
